Make default battery selection more predictable.
In the past, the system simply picked a battery from
the available batteries. An alphabetic search meant it
was deterministic, and because of how DeviceInstances
are allocated it more or less corresponded with the
lowest device instance, and canbus BMSes were prioritised.

With this change, the BMS with the lowest device instance
is explicitly prioritised over non-BMS monitors (eg BMV,
smartshunt). If no BMS in the system, then use the battery
monitor with the lowest DeviceInstance.

https://github.com/victronenergy/venus/issues/901