display external control state w/ managed battery
This used to be implemented in mk2-dbus, but it turns out that
the information in com.victronenergy.vebus/State carries useful
information that is now lost when it is simply set to External
Control with a managed battery. Since this was meant to avoid
confusion about the state shown on the Overviews, it is
sufficient to move that to the SystemState instead.