dess: consider acpv on output if charge restricted
When charging from the grid is restricted, the system is in
self-consumption strategy, and there is AC-coupled PV on the
output this was sold to the grid instead of used for charging.
This bug crept in in Venus 3.30 when switching (back) to using
DC charge current control to control charging instead of the AC
setpoint (since that caused full speed charging in many cases).