scheduled charging: don't keep writing overrides
The override paths for charging and preventing battery discharge
are written every 5 seconds, unless another delegate (DynamicEss)
has control. This makes it impossible to use those overrides
for custom work, eg via Node-Red.

Now, we only clear the overrides if we have previously acquired
control of the charge machinery.

This could potentially cause stuck values after a restart of
systemcalc, but hub4control has its own timeout built in, and
after a reboot all overrides are removed anyway. So this should
be alright.